Prep for v0.14.0 - reset and modernize repo and tooling (#1)
This is the start of a 0.14.0 release. This PR does not fix any of the deprecation warnings that we've been seeing, it just resets and modernises the tooling in the repo. When this PR is merged, I can create another branch that specifically fixes only the Ruby 2.7 deprecations, so they can be handled in isolation.

This PR contains the following changes:

* Update the Gemspec to reflect the fact that it's our repo/Gem
* Update the Gem's dev dependencies
* Update the Ruby version to match Panacea
* Reset the Rubocop config with up-to-date rules
* Applies Rubocop autocorrections to lib and spec code
* Updates RSpec config
* Adds Simplecov
* Adds Dependabot config
* Adds RSpec and Linters CI actions
* Removes Travis CI and Codeclimate config
* Adds Guard and config
* Adds generated binstubs

For the purposes of clarity - anything changed in lib or spec was more than likely generated by the updated Rubocop rules. You can probably ignore the changes in those directories. Any changes to the specs have changed the formatting only, and not the actual code under test.

Remarkably, the code coverage of the lib directory is 98.16%, so given that the test suite still passes with no new or modified specs, I'm confident that the changes Rubocop has made to the code have not changed any functionality.

These changes are mostly about making this repo easier to work with. When this has merged we can focus on the deprecation warning fixes, and THEN tag and release the 0.14.0 version for installation in Panacea.